Group ingredients and add measurements
When reading the recipe on a mobile, it is more convenient if you
do not have to scroll.  Since items like potatoes and bell peppers
are usually prepared and put into bowls, their measurements do
not need to be included in the text.  By putting  measurements
for things like fish sauce in the text, I hope to minimize
the need to scroll.